test: rename fixture
