+++ packages/x-adapter-platform/src/__tests__/__fixtures__/recommendations.response.ts
@@ -1,4 +1,4 @@
-import { platformResult } from './result.response'
+import { platformResult } from './result'
 
 export const platformRecommendationsResponse = {
   topclicked: {
