+++ packages/x-adapter-platform/src/__tests__/__fixtures__/result.ts
