+++ packages/x-adapter-platform/src/mappers/responses/__tests__/recommendations-response.mapper.spec.ts
@@ -1,7 +1,7 @@
 import type { PlatformRecommendationsResponse } from '../../../types/responses/recommendations-response.model'
 
 import { platformRecommendationsResponse } from '../../../__tests__/__fixtures__/recommendations.response'
-import { result } from '../../../__tests__/__fixtures__/result.response'
+import { result } from '../../../__tests__/__fixtures__/result'
 import { recommendationsResponseMapper } from '../recommendations-response.mapper'
 
 describe('recommendationsResponseMapper tests', () => {
